requirements.txt
awkward
click
coffea==2025.1.1
colorama
dask-awkward==2025.5.0
hist
lz4
numpy
pandas
PyYAML
scikit-learn
scipy
vector
xgboost>=2.0.3
XRootD
zstandard
rhalphalib